insta-steel amount is in pixels

+++ CrisisAtSwissStation/GameEngine.cs
@@ -268,7 +268,7 @@ namespace CrisisAtSwissStation
 
                 if (parseWin)
                 {
-                     ScrollingWorld.numDrawLeft = Int32.Parse(temp);
+                     ScrollingWorld.numDrawLeft = (float) Int32.Parse(temp);
                 }
             }
 
@@ -313,7 +313,7 @@ namespace CrisisAtSwissStation
             //Draw IS label 
             spriteBatch.DrawString(spriteFont, "Insta-Steel :", new Vector2(50, 140), Color.Red);
             //Draw IS amount
-            spriteBatch.DrawString(spriteFont, ScrollingWorld.numDrawLeft.ToString(), new Vector2(155, 140), Color.Yellow);
+            spriteBatch.DrawString(spriteFont, ((int)ScrollingWorld.numDrawLeft).ToString(), new Vector2(155, 140), Color.Yellow);
 
             //Draw Jump label 
             spriteBatch.DrawString(spriteFont, "Jump :", new Vector2(50, 200), Color.Red);
